Use Uhura for generating update.rdf
Instead of doing some sed magic, relying on The Cloud and HTTPS sign
packages in the standard Mozilla manner. Uhura seems to be the most
appropriate tool for it, fetch it from
http://www.softlights.net/projects/mxtools/

This patch requires an makexpi.local file with local definitions for
where to find Uhura, the key file and the update URL, but there's an
example file.