Refactor build output
We have to link to a specific version in update.rdf, not just to
latest.xpi, because firefox seems to fetch the update.rdf only once and
when this xpi is changed before the user chooses to update the hash is
wrong.

So for now put versioned .xpis into pkg/, make a symlink to latest.xpi
there and also put the update.rdf there that links to the latest xpi
(not: latest.xpi). The whole pkg/ directory can then be rsync'd to a web
server.